requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
werkzeug==3.0.1 